requirements.txt





celery==5.2.3


dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==2.2.28
django-allauth==0.37.1
django-cors-headers==3.11.0
django-forms-bootstrap==3.1.0
django-phonenumber-field==6.0.0
django-rest-auth==0.9.3
django-rest-passwordreset==1.2.1
django-ses==2.3.1
django-settings-export==1.2.1
django-storages==1.12.3
django-timezone-field==5.0
django-tinymce==3.4.0
djangoflutterwave==0.3.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
dramatiq==1.13.0
drf-compound-fields==2.0.0
drf-yasg==1.20.0

fastjsonschema==2.15.3

geocoder==1.38.1
gevent==21.12.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0

gunicorn==19.7.1
h5py==3.6.0
html5lib==1.1




oauthlib==3.2.0


postgres==4.0

psycopg2==2.9.3
psycopg2-binary==2.9.3
psycopg2-pool==1.1

PyJWT==1.7.1
pyparsing==3.0.8

pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python3-openid==3.1.0

redis==4.1.1
regex==2022.7.9
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8


sqlparse==0.3.0


tinycss2==1.1.1


urllib3==1.26.10
uuid==1.30
vine==5.0.0
virtualenv==20.14.1
whitenoise==5.3.0
wrapt==1.14.0